Update webp-pixbuf-loader
Fix segfault and crash when copying image into clipboard (see pull
request #10).

The project is currently ported to meson, but I couldn't get it to
build, so I used the latest commit that still used cmake-ninja. There
doesn't seem to be any bugs fixed since it was ported, so I don't think
we are missing anything.